refactor: 重构任务提示
